feat/cleanup/docs: sistema de saldo financeiro, correções de bugs e documentação completa
## feat: novo sistema de pontuação financeira

- Substituído modelo de pontos inteiros (0/1/3) por saldo monetário em R$
  - Quem acerta o placar exato recebe +R$ valor_fase
  - Quem erra (mas alguém acertou) paga -R$ valor_fase
  - Se ninguém acertar, o jogo é neutro (R$ 0,00 para todos)
- `Aposta.pontos` alterado de `Integer` para `Numeric(10, 2)` para suportar valores decimais
- `recalcular_apostas` reescrito em `service/jogos.py` com lógica win/loss baseada no `valor` da fase
- `calcular_placar` reescrito em `service/placar.py`:
  - Campos renomeados: `total_pontos` → `saldo_total`, `pontos` → `saldo`
  - Removido `total_gasto` (não utilizado no novo modelo)
  - Guard `float(a.pontos or 0)` para apostas com pontos NULL legados
- Schemas Pydantic em `esquemas.py` atualizados: `PlacarFase.pontos: int` → `saldo: float`,
  `PlacarParticipante` sem `total_gasto`, `ApostaSaida.pontos: float`
- Controller `placar.py` atualizado para mapear os novos campos

## fix: bugs críticos em service/apostas.py

- `criar_aposta`: corrigido `not (jogo or jogo.encerrado)` → `not jogo or jogo.encerrado`
  (a versão com parênteses causava AttributeError ao acessar `None.encerrado`)
- `deletar_aposta`: mesma correção de lógica booleana
- `service/placar.py`: guard defensivo `or 0` em `float(a.pontos or 0)` para evitar
  TypeError com pontos NULL no banco

## fix: placar não carregava no frontend

- `app.js`: `f.saldo ?? 0` para evitar `TypeError: undefined.toFixed()` quando saldo é null
- `index.html`: adicionado `?v=2` no script tag para forçar cache busting após atualização do JS
  (o browser mantinha cache do `app.js` antigo com `total_gasto` que não existe mais na API)

## cleanup: cosmético no backend

- `modulo.py`: linha em branco entre decoradores e definições de função (estilo PEP 8)
- `popular.py`: newline no final do arquivo
- `controller/jogos.py`, `controller/apostas.py`: padronização de condicionais

## docs: README reescrito com documentação completa

- Badges de tecnologias (Python, FastAPI, SQLAlchemy, Pydantic, PostgreSQL, Docker, Nginx, HTML/CSS/JS)
- Árvore de arquitetura do projeto com badges inline
- Tabela de modelos de dados com descrição de cada entidade
- Tabela completa de endpoints da API REST com métodos, rotas e descrições
- Explicação detalhada do sistema de pontuação financeira (tabela de valores por fase)
- Seção "Como Rodar" com Docker Compose: comandos de start, stop, reset de volume
- Comandos Docker úteis: rebuild sem cache, logs, shell no container, prune
- Tabela de versões das dependências principais
- Descrição das camadas Controller / Service / Schemas / Models

+++ backend/app/modelos.py
@@ -79,7 +79,7 @@ class Aposta(Base):
     palpite_casa = Column(Integer, nullable=False)
     palpite_fora = Column(Integer, nullable=False)
     valor = Column(Numeric(10, 2), nullable=False, default=1)
-    pontos = Column(Integer, default=0, nullable=False)
+    pontos = Column(Numeric(10, 2), default=0, nullable=False)
 
     __table_args__ = (UniqueConstraint("id_participante", "id_jogo", name="uq_aposta_participante_jogo"),)
